swagger setup done

+++ mobile_project/mobile_sale/views.py
@@ -1,25 +1,16 @@
-from rest_framework.exceptions import APIException
-from rest_framework.generics import GenericAPIView
-from django.db import IntegrityError
-from django.contrib.auth import authenticate
 from rest_framework.views import APIView
 from rest_framework.response import Response
 from rest_framework import status
 from django.shortcuts import get_object_or_404
+from django.contrib.auth import authenticate
 from .models import Product, Reviews, Inventory, Order
-from .serializers import ProductSerializer, ReviewSerializer, InventorySerializer, OrderSerializer, RegisterSerializer, LoginSerializer
-from rest_framework.authtoken.models import Token
-from django.contrib.auth.models import User
-from .serializers import UserSerializer
-from django.db import IntegrityError
+from .serializers import ProductSerializer, ReviewSerializer, InventorySerializer, OrderSerializer, RegisterSerializer, LoginSerializer, UserSerializer
 from rest_framework.permissions import IsAuthenticated
+from rest_framework.authtoken.models import Token
 from rest_framework.authentication import TokenAuthentication
+from rest_framework.generics import GenericAPIView
 
-class ProductView(GenericAPIView):
-    # authentication_classes = [TokenAuthentication]
-
-    # permission_classes = [IsAuthenticated]
-    serializer_class = ProductSerializer
+class ProductAPIView(APIView):
     def get(self, request, pk=None):
         if pk:
             product = get_object_or_404(Product, pk=pk)
@@ -29,45 +20,32 @@ class ProductView(GenericAPIView):
             serializer = ProductSerializer(products, many=True)
         return Response(serializer.data)
 
+    def put(self, request, pk=None):
+        if not pk:
+            return Response({"message": "ID is required for update"}, status=status.HTTP_400_BAD_REQUEST)
+        product = get_object_or_404(Product, pk=pk)
+        serializer = ProductSerializer(product, data=request.data, partial=True)
+        if serializer.is_valid():
+            serializer.save()
+            return Response(serializer.data)
+        return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
+
+    def delete(self, request, pk=None):
+        if not pk:
+            return Response({"message": "ID is required for delete"}, status=status.HTTP_400_BAD_REQUEST)
+        product = get_object_or_404(Product, pk=pk)
+        product.delete()
+        return Response({"success": f"Product with ID {pk} has been deleted."}, status=status.HTTP_200_OK)
+
+class ProductCreateAPIView(APIView):
     def post(self, request):
         serializer = ProductSerializer(data=request.data)
         if serializer.is_valid():
-            try:
-                serializer.save()
-                return Response(serializer.data, status=status.HTTP_201_CREATED)
-            except IntegrityError:
-                return Response({"error": "A product with this name already exists."}, status=status.HTTP_400_BAD_REQUEST)
+            serializer.save()
+            return Response(serializer.data, status=status.HTTP_201_CREATED)
         return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
 
-    def put(self, request):
-        product_id = request.data.get("id")
-        if not product_id:
-            return Response({"error": "Product ID is required in the JSON body"}, status=status.HTTP_400_BAD_REQUEST)
-
-        product = get_object_or_404(Product, pk=product_id)
-        serializer = ProductSerializer(product, data=request.data, partial=True)
-        if serializer.is_valid():
-            try:
-                serializer.save()
-                return Response(serializer.data)
-            except IntegrityError:
-                return Response({"error": "A product with this name already exists."}, status=status.HTTP_400_BAD_REQUEST)
-        return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
-
-    def delete(self, request):
-        product_id = request.data.get("id")
-        if not product_id:
-            return Response({"error": "Product ID is required in the JSON body"}, status=status.HTTP_400_BAD_REQUEST)
-
-        product = get_object_or_404(Product, pk=product_id)
-        product.delete()
-        return Response({"success": f"Product with ID {product_id} has been deleted."}, status=status.HTTP_200_OK)
-
-class ReviewView(GenericAPIView):
-    # authentication_classes = [TokenAuthentication]
-    # permission_classes = [IsAuthenticated]
-    serializer_class = ReviewSerializer
-
+class ReviewAPIView(APIView):
     def get(self, request, pk=None):
         if pk:
             review = get_object_or_404(Reviews, pk=pk)
@@ -77,15 +55,10 @@ class ReviewView(GenericAPIView):
             serializer = ReviewSerializer(reviews, many=True)
         return Response(serializer.data)
 
-    def post(self, request):
-        serializer = ReviewSerializer(data=request.data)
-        if serializer.is_valid():
-            serializer.save(user=request.user)
-            return Response(serializer.data, status=status.HTTP_201_CREATED)
-        return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
-
     def put(self, request, pk=None):
-        review = get_object_or_404(Reviews, pk=request.data.get("id"))
+        if not pk:
+            return Response({"message": "ID is required for update"}, status=status.HTTP_400_BAD_REQUEST)
+        review = get_object_or_404(Reviews, pk=pk)
         serializer = ReviewSerializer(review, data=request.data, partial=True)
         if serializer.is_valid():
             serializer.save()
@@ -93,15 +66,21 @@ class ReviewView(GenericAPIView):
         return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
 
     def delete(self, request, pk=None):
-        review = get_object_or_404(Reviews, pk=request.data.get("id"))
+        if not pk:
+            return Response({"message": "ID is required for delete"}, status=status.HTTP_400_BAD_REQUEST)
+        review = get_object_or_404(Reviews, pk=pk)
         review.delete()
         return Response({"success": "Review deleted."}, status=status.HTTP_200_OK)
 
-class InventoryView(GenericAPIView):
-    # authentication_classes = [TokenAuthentication]
-    # permission_classes = [IsAuthenticated]
-    serializer_class = InventorySerializer
+class ReviewCreateAPIView(APIView):
+    def post(self, request):
+        serializer = ReviewSerializer(data=request.data)
+        if serializer.is_valid():
+            serializer.save(user=request.user)
+            return Response(serializer.data, status=status.HTTP_201_CREATED)
+        return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
 
+class InventoryAPIView(APIView):
     def get(self, request, pk=None):
         if pk:
             inventory = get_object_or_404(Inventory, pk=pk)
@@ -111,19 +90,10 @@ class InventoryView(GenericAPIView):
             serializer = InventorySerializer(inventories, many=True)
         return Response(serializer.data)
 
-    def post(self, request):
-        serializer = InventorySerializer(data=request.data)
-        if serializer.is_valid():
-            serializer.save()
-            return Response(serializer.data, status=status.HTTP_201_CREATED)
-        return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
-
     def put(self, request, pk=None):
-        inventory_id = request.data.get("id")
-        if not inventory_id:
-            return Response({"error": "Inventory ID is required in the JSON body"}, status=status.HTTP_400_BAD_REQUEST)
-
-        inventory = get_object_or_404(Inventory, pk=inventory_id)
+        if not pk:
+            return Response({"message": "ID is required for update"}, status=status.HTTP_400_BAD_REQUEST)
+        inventory = get_object_or_404(Inventory, pk=pk)
         serializer = InventorySerializer(inventory, data=request.data, partial=True)
         if serializer.is_valid():
             serializer.save()
@@ -131,21 +101,23 @@ class InventoryView(GenericAPIView):
         return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
 
     def delete(self, request, pk=None):
-        inventory_id = request.data.get("id")
-        if not inventory_id:
-            return Response({"error": "Inventory ID is required in the JSON body"}, status=status.HTTP_400_BAD_REQUEST)
-
-        inventory = get_object_or_404(Inventory, pk=inventory_id)
+        if not pk:
+            return Response({"message": "ID is required for delete"}, status=status.HTTP_400_BAD_REQUEST)
+        inventory = get_object_or_404(Inventory, pk=pk)
         inventory.delete()
-        return Response({"success": f"Inventory with ID {inventory_id} has been deleted."}, status=status.HTTP_200_OK)
+        return Response({"success": f"Inventory with ID {pk} has been deleted."}, status=status.HTTP_200_OK)
 
-from rest_framework.permissions import IsAuthenticated
-from rest_framework.authentication import TokenAuthentication
+class InventoryCreateAPIView(APIView):
+    def post(self, request):
+        serializer = InventorySerializer(data=request.data)
+        if serializer.is_valid():
+            serializer.save()
+            return Response(serializer.data, status=status.HTTP_201_CREATED)
+        return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
 
-class OrderView(GenericAPIView):
+class OrderAPIView(APIView):
     authentication_classes = [TokenAuthentication]
     permission_classes = [IsAuthenticated]
-    serializer_class = OrderSerializer
 
     def get(self, request, pk=None):
         if pk:
@@ -156,15 +128,10 @@ class OrderView(GenericAPIView):
             serializer = OrderSerializer(orders, many=True)
         return Response(serializer.data)
 
-    def post(self, request):
-        serializer = OrderSerializer(data=request.data)
-        if serializer.is_valid():
-            serializer.save(user=request.user)
-            return Response(serializer.data, status=status.HTTP_201_CREATED)
-        return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
-
     def put(self, request, pk=None):
-        order = get_object_or_404(Order, pk=request.data.get("id"), user=request.user)
+        if not pk:
+            return Response({"message": "ID is required for update"}, status=status.HTTP_400_BAD_REQUEST)
+        order = get_object_or_404(Order, pk=pk, user=request.user)
         serializer = OrderSerializer(order, data=request.data, partial=True)
         if serializer.is_valid():
             serializer.save()
@@ -172,50 +139,51 @@ class OrderView(GenericAPIView):
         return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
 
     def delete(self, request, pk=None):
-        order = get_object_or_404(Order, pk=request.data.get("id"), user=request.user)
+        if not pk:
+            return Response({"message": "ID is required for delete"}, status=status.HTTP_400_BAD_REQUEST)
+        order = get_object_or_404(Order, pk=pk, user=request.user)
         order.delete()
         return Response({"success": "Order deleted."}, status=status.HTTP_200_OK)
 
+class OrderCreateAPIView(APIView):
+    authentication_classes = [TokenAuthentication]
+    permission_classes = [IsAuthenticated]
+
+    def post(self, request):
+        serializer = OrderSerializer(data=request.data)
+        if serializer.is_valid():
+            serializer.save(user=request.user)
+            return Response(serializer.data, status=status.HTTP_201_CREATED)
+        return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
+
 class RegisterAPI(GenericAPIView):
     serializer_class = RegisterSerializer
+
     def post(self, request):
         data = request.data
         serializer = RegisterSerializer(data=data)
         
         if not serializer.is_valid():
-            return Response({
-                'message': serializer.errors
-            }, status=status.HTTP_400_BAD_REQUEST)
+            return Response({'message': serializer.errors}, status=status.HTTP_400_BAD_REQUEST)
         
         user = serializer.save()
         token, created = Token.objects.get_or_create(user=user)
-        return Response({
-            'token': token.key,
-            'user': UserSerializer(user).data
-        }, status=status.HTTP_201_CREATED)
+        return Response({'token': token.key, 'user': UserSerializer(user).data}, status=status.HTTP_201_CREATED)
 
 class LoginAPI(GenericAPIView):
     serializer_class = LoginSerializer
+
     def post(self, request):
         data = request.data
         serializer = LoginSerializer(data=data)
         
         if not serializer.is_valid():
-            return Response({
-                'message': serializer.errors
-            }, status=status.HTTP_400_BAD_REQUEST)
+            return Response({'message': serializer.errors}, status=status.HTTP_400_BAD_REQUEST)
         
         user = authenticate(username=serializer.validated_data['username'], password=serializer.validated_data['password'])
         
         if not user:
-            return Response({
-                'message': 'Invalid credentials'
-            }, status=status.HTTP_401_UNAUTHORIZED)
+            return Response({'message': 'Invalid credentials'}, status=status.HTTP_401_UNAUTHORIZED)
         
         token, created = Token.objects.get_or_create(user=user)
-        return Response({
-            'token': token.key,
-            'user': UserSerializer(user).data
-        }, status=status.HTTP_200_OK)
-
-
+        return Response({'token': token.key, 'user': UserSerializer(user).data}, status=status.HTTP_200_OK)
